mmap: add suport for /dev/mem
During the creation of a GuestRegionMmap for the VM memory zone, the MmapRegion::build function from vm-memory is invoked. In this context, the check_file_offset function is employed to verify the file size Notably, an operation involving seeking from the end of the file is carried out. It is important to note that for the Linux /dev/mem handled by drivers/char/mem.c, which implements the character device, the seek file operation is not supported. This is attributed to the fact that in memory devices, the full 32/64 bits of the offset are utilized, rendering the seek operation meaningless.

The provided patch serves as a provisional resolution that circumvents that specific corner.

Signed-off-by: joaopeixoto13 <joaopeixotooficial@gmail.com>